+++ src/Translatable/Query/TreeWalker/TranslationWalker.php
@@ -2,8 +2,9 @@
 
 namespace Gedmo\Translatable\Query\TreeWalker;
 
-use Doctrine\DBAL\Platforms\MySqlPlatform;
-use Doctrine\DBAL\Platforms\PostgreSqlPlatform;
+use Doctrine\DBAL\Platforms\MySQLPlatform;
+use Doctrine\DBAL\Platforms\PostgreSQL94Platform;
+use Doctrine\DBAL\Platforms\PostgreSQLPlatform;
 use Doctrine\DBAL\Types\Type;
 use Doctrine\ORM\Mapping\ClassMetadata;
 use Doctrine\ORM\Query;
@@ -346,9 +347,9 @@ class TranslationWalker extends SqlWalker
 
                 // Treat translation as original field type
                 $fieldMapping = $meta->getFieldMapping($field);
-                if ((($this->platform instanceof MySqlPlatform) &&
+                if ((($this->platform instanceof MySQLPlatform) &&
                     in_array($fieldMapping['type'], ['decimal'])) ||
-                    (!($this->platform instanceof MySqlPlatform) &&
+                    (!($this->platform instanceof MySQLPlatform) &&
                     !in_array($fieldMapping['type'], ['datetime', 'datetimetz', 'date', 'time']))) {
                     $type = Type::getType($fieldMapping['type']);
                     $substituteField = 'CAST('.$substituteField.' AS '.$type->getSQLDeclaration($fieldMapping, $this->platform).')';
@@ -460,8 +461,12 @@ class TranslationWalker extends SqlWalker
             return $component;
         }
 
+        // @todo: remove the `PostgreSQL94Platform` check when dropping doctrine/dbal 3.1.x support
+        // the below check prefers the `PostgreSQLPlatform` class for doctrine/dbal 2.13.x or 3.2
+        // and later and falls back to the `PostgreSQL94Platform` class for compatibility with 3.1
+        // where `PostgreSQLPlatform` does not exist
         // try to look at postgres casting
-        if ($this->platform instanceof PostgreSqlPlatform) {
+        if ($this->platform instanceof PostgreSQLPlatform || $this->platform instanceof PostgreSQL94Platform) {
             switch ($typeFK) {
                 case 'string':
                 case 'guid':